Move MainContext from CreateCVAPP to index.
