+++ frontend/src/components/pdf/templates/upeksa/Upeksa.js
@@ -8,7 +8,7 @@ import Details from '../../Details'
 import Languages from '../../Languages'
 import Hobbies from '../../Hobbies'
 import GDPA from '../../GDPA'
-import MainContext from '../../../../CreateCVApp'
+import MainContext from '../../../../index'
 import { Provider } from 'react-redux'
 import store from '../../../../redux/store'
 import Spacer from '../../Spacer'
