+++ frontend/src/components/ui/createcv/Links.js
@@ -12,7 +12,7 @@ import {
     CardEmpty
 } from '@salesforce/design-system-react';
 import { connect } from "react-redux"
-import MainContext from '../../../CreateCVApp'
+import MainContext from '../../../index'
 import debounce from '../../../utilities/debounce'
 import { updatePreview } from '../../../redux/reducers/pdf/pdfViewer/actions'
 import { debounceTime } from '../../../utilities/variables'
